Api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.5.14
datasets==2.14.5
dill==0.3.7
docopt==0.6.2
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.305
langsmith==0.0.41
MarkupSafe==2.1.3
marshmallow==3.20.1
minio==7.1.17
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
onnxruntime==1.16.0
openai==0.28.1
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pipreqs==0.4.13
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pyarrow==13.0.0
pydantic==1.10.13
pymilvus==2.3.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
pymilvus==2.3.1
fastapi==0.103.2
uvicorn==0.23.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
